+++ src/Neo4jDFG/Neo4jDFG.jl
@@ -1,4 +1,4 @@
-module CloudGraphsDFGs
+module Neo4jDFGs
 
 using ...DistributedFactorGraphs
 
@@ -105,16 +105,16 @@ using SparseArrays
 using UUIDs
 
 # Entities
-include("entities/CloudGraphsDFG.jl")
+include("entities/Neo4jDFG.jl")
 include("entities/CGStructure.jl")
 
 # Services
 include("services/CommonFunctions.jl")
 include("services/CGStructure.jl")
-include("services/CloudGraphsDFG.jl")
+include("services/Neo4jDFG.jl")
 
 # Exports
-export Neo4jInstance, CloudGraphsDFG
+export Neo4jInstance, Neo4jDFG
 
 # Additional exports for CGStructure
 export copySession!
@@ -126,5 +126,8 @@ export existsSession, existsRobot, existsUser
 export getSession, getRobot, getUser
 export updateSession, updateRobot, updateUser
 export lsSessions, lsRobots, lsUsers
+
+const CloudGraphsDFG = Neo4jDFG
+# const CloudGraphsDFG{T} = Neo4jDFG{T}
 
 end