+++ src/Neo4jDFG/services/CGStructure.jl
@@ -1,11 +1,13 @@
-function isValidLabel(abstractNode::N)::Bool where N <: AbstractCGNode
+# return ::Bool
+function isValidLabel(abstractNode::N) where N <: AbstractCGNode
     id = String(abstractNode.id)
     return all(t -> t != uppercase(id), _invalidIds) && match(_validLabelRegex, id) !== nothing
 end
 
 # Fastest way I can think to convert the data into a dict
 #TODO: Probably should be made more efficient...definitely should be made more efficient
-function _convertNodeToDict(abstractNode::N)::Dict{String, Any} where N <: AbstractCGNode
+#  return ::Dict{String, Any}
+function _convertNodeToDict(abstractNode::N) where N <: AbstractCGNode
     cp = deepcopy(abstractNode)
     data = length(cp.data) != 0 ? JSON2.write(cp.data) : "{}"
     ser = JSON2.read(JSON2.write(abstractNode), Dict{String, Any})
@@ -15,7 +17,8 @@ function _convertNodeToDict(abstractNode::N)::Dict{String, Any} where N <: Abstr
 end
 
 #TODO: Refactor, #HACK :D (but it works!)
-function _convertDictToSession(dict::Dict{String, Any})::Session
+# return ::Session
+function _convertDictToSession(dict::Dict{String, Any})
     data = JSON2.read(String(base64decode(dict["data"])), Dict{Symbol, String})
     session = Session(
         Symbol(dict["id"]),
@@ -29,7 +32,8 @@ function _convertDictToSession(dict::Dict{String, Any})::Session
     return session
 end
 #TODO: Refactor, #HACK :D (but it works!)
-function _convertDictToRobot(dict::Dict{String, Any})::Robot
+# returns ::Robot
+function _convertDictToRobot(dict::Dict{String, Any})
     data = JSON2.read(String(base64decode(dict["data"])), Dict{Symbol, String})
     robot = Robot(
         Symbol(dict["id"]),
@@ -42,7 +46,8 @@ function _convertDictToRobot(dict::Dict{String, Any})::Robot
     return robot
 end
 #TODO: Refactor, #HACK :D (but it works!)
-function _convertDictToUser(dict::Dict{String, Any})::User
+# returns ::User
+function _convertDictToUser(dict::Dict{String, Any})
     data = JSON2.read(String(base64decode(dict["data"])), Dict{Symbol, String})
     user = User(
         Symbol(dict["id"]),
@@ -54,7 +59,8 @@ function _convertDictToUser(dict::Dict{String, Any})::User
     return user
 end
 
-function createUser(dfg::CloudGraphsDFG, user::User)::User
+# returns ::User
+function createUser(dfg::Neo4jDFG, user::User)
     Symbol(dfg.userId) != user.id && error("DFG user ID must match user's ID")
     !isValidLabel(user) && error("Node cannot have an ID '$(user.id)'.")
 
@@ -64,7 +70,8 @@ function createUser(dfg::CloudGraphsDFG, user::User)::User
     return user
 end
 
-function createRobot(dfg::CloudGraphsDFG, robot::Robot)::Robot
+# returns ::Robot
+function createRobot(dfg::Neo4jDFG, robot::Robot)
     Symbol(dfg.robotId) != robot.id && error("DFG robot ID must match robot's ID")
     Symbol(dfg.userId) != robot.userId && error("DFG user ID must match robot's user ID")
     !isValidLabel(robot) && error("Node cannot have an ID '$(robot.id)'.")
@@ -84,7 +91,8 @@ function createRobot(dfg::CloudGraphsDFG, robot::Robot)::Robot
     return robot
 end
 
-function createSession(dfg::CloudGraphsDFG, session::Session)::Session
+# returns ::Session
+function createSession(dfg::Neo4jDFG, session::Session)
     Symbol(dfg.robotId) != session.robotId && error("DFG robot ID must match session's robot ID")
     Symbol(dfg.userId) != session.userId && error("DFG user ID must match session's->robot's->user ID")
     !isValidLabel(session) && error("Node cannot have an ID '$(session.id)'.")
@@ -107,8 +115,11 @@ end
 """
 $(SIGNATURES)
 Shortcut method to create the user, robot, and session if it doesn't already exist.
+
+Notes
+- return `::Session`
 """
-function createDfgSessionIfNotExist(dfg::CloudGraphsDFG)::Session
+function createDfgSessionIfNotExist(dfg::Neo4jDFG)
     strip(dfg.userId) == "" && error("User ID is not populated in DFG.")
     strip(dfg.robotId) == "" && error("Robot ID is not populated in DFG.")
     strip(dfg.sessionId) == "" && error("Session ID is not populated in DFG.")
@@ -129,8 +140,11 @@ end
 $(SIGNATURES)
 List all sessions for the specified DFG's robot and user.
 Returns nothing if it isn't found.
+
+Notes
+- Returns `Vector{Session}`
 """
-function lsSessions(dfg::CloudGraphsDFG)::Vector{Session}
+function lsSessions(dfg::Neo4jDFG)
     sessionNodes = _getNeoNodesFromCyphonQuery(dfg.neo4jInstance, "(node:SESSION:$(dfg.robotId):$(dfg.userId))")
     return map(s -> _convertDictToSession(Neo4j.getnodeproperties(s)), sessionNodes)
 end
@@ -139,8 +153,11 @@ end
 $(SIGNATURES)
 List all robots for the specified DFG's user.
 Returns nothing if it isn't found.
+
+Notes
+- Returns `::Vector{Robot}`
 """
-function lsRobots(dfg::CloudGraphsDFG)::Vector{Robot}
+function lsRobots(dfg::Neo4jDFG)
     robotNodes = _getNeoNodesFromCyphonQuery(dfg.neo4jInstance, "(node:ROBOT:$(dfg.userId))")
     return map(s -> _convertDictToRobot(Neo4j.getnodeproperties(s)), robotNodes)
 end
@@ -149,8 +166,11 @@ end
 $(SIGNATURES)
 List all users.
 Returns nothing if it isn't found.
+
+Notes
+- Returns `::Vector{User}`
 """
-function lsUsers(dfg::CloudGraphsDFG)::Vector{User}
+function lsUsers(dfg::Neo4jDFG)
     userNodes = _getNeoNodesFromCyphonQuery(dfg.neo4jInstance, "(node:USER)")
     return map(s -> _convertDictToUser(Neo4j.getnodeproperties(s)), userNodes)
 end
@@ -159,8 +179,11 @@ end
 $(SIGNATURES)
 Get a session specified by userId:robotId:sessionId.
 Returns nothing if it isn't found.
+
+Notes
+- Returns either `::Union{Session, Nothing}`
 """
-function getSession(dfg::CloudGraphsDFG, userId::Symbol, robotId::Symbol, sessionId::Symbol)::Union{Session, Nothing}
+function getSession(dfg::Neo4jDFG, userId::Symbol, robotId::Symbol, sessionId::Symbol)
     !isValidLabel(userId) && error("Can't retrieve session with user ID '$(userId)'.")
     !isValidLabel(robotId) && error("Can't retrieve session with robot ID '$(robotId)'.")
     !isValidLabel(sessionId) && error("Can't retrieve session with session ID '$(sessionId)'.")
@@ -174,8 +197,11 @@ end
 $(SIGNATURES)
 Get the session specified by the DFG object.
 Returns nothing if it isn't found.
+
+Notes
+- Returns either `::Union{Nothing, Session}`
 """
-function getSession(dfg::CloudGraphsDFG)::Union{Nothing, Session}
+function getSession(dfg::Neo4jDFG)
     return getSession(dfg, Symbol(dfg.userId), Symbol(dfg.robotId), Symbol(dfg.sessionId))
 end
 
@@ -183,8 +209,11 @@ end
 $(SIGNATURES)
 Get a robot specified by userId:robotId.
 Returns nothing if it isn't found.
+
+Notes
+- Returns either `::Union{Robot, Nothing}`
 """
-function getRobot(dfg::CloudGraphsDFG, userId::Symbol, robotId::Symbol)::Union{Robot, Nothing}
+function getRobot(dfg::Neo4jDFG, userId::Symbol, robotId::Symbol)
     !isValidLabel(userId) && error("Can't retrieve robot with user ID '$(userId)'.")
     !isValidLabel(robotId) && error("Can't retrieve robot with robot ID '$(robotId)'.")
     robotNode = _getNeoNodesFromCyphonQuery(dfg.neo4jInstance, "(node:ROBOT:$(robotId):$(userId))")
@@ -197,8 +226,11 @@ end
 $(SIGNATURES)
 Get the robot specified by the DFG object.
 Returns nothing if it isn't found.
+
+Notes
+- Returns either `::Union{Nothing, Robot}`
 """
-function getRobot(dfg::CloudGraphsDFG)::Union{Nothing, Robot}
+function getRobot(dfg::Neo4jDFG)
     return getRobot(dfg, Symbol(dfg.userId), Symbol(dfg.robotId))
 end
 
@@ -206,8 +238,11 @@ end
 $(SIGNATURES)
 Get a user specified by userId.
 Returns nothing if it isn't found.
+
+Notes
+- Returns either `::Union{User, Nothing}`
 """
-function getUser(dfg::CloudGraphsDFG, userId::Symbol)::Union{User, Nothing}
+function getUser(dfg::Neo4jDFG, userId::Symbol)
     !isValidLabel(userId) && error("Can't retrieve user with user ID '$(userId)'.")
     userNode = _getNeoNodesFromCyphonQuery(dfg.neo4jInstance, "(node:USER:$(userId))")
     length(userNode) == 0 && return nothing
@@ -219,8 +254,11 @@ end
 $(SIGNATURES)
 Get the user specified by the DFG object.
 Returns nothing if it isn't found.
+
+Notes
+- Returns either `::Union{Nothing, User}`
 """
-function getUser(dfg::CloudGraphsDFG)::Union{Nothing, User}
+function getUser(dfg::Neo4jDFG)
     return getUser(dfg, Symbol(dfg.userId))
 end
 
@@ -228,8 +266,11 @@ end
 """
     $(SIGNATURES)
 DANGER: Clears the whole session from the database.
+
+Notes
+- Returns `::Nothing`
 """
-function clearSession!!(dfg::CloudGraphsDFG)::Nothing
+function clearSession!!(dfg::Neo4jDFG)
     # Perform detach+deletion
     _queryNeo4j(dfg.neo4jInstance, "match (node:$(dfg.userId):$(dfg.robotId):$(dfg.sessionId)) detach delete node ")
 
@@ -241,8 +282,11 @@ end
 """
     $(SIGNATURES)
 DANGER: Clears the whole robot + sessions from the database.
+
+Notes
+- Returns `::Nothing`
 """
-function clearRobot!!(dfg::CloudGraphsDFG)::Nothing
+function clearRobot!!(dfg::Neo4jDFG)
     # Perform detach+deletion
     _queryNeo4j(dfg.neo4jInstance, "match (node:$(dfg.userId):$(dfg.robotId)) detach delete node ")
 
@@ -254,8 +298,11 @@ end
 """
     $(SIGNATURES)
 DANGER: Clears the whole user + robot + sessions from the database.
+
+Notes
+- Returns `::Nothing`
 """
-function clearUser!!(dfg::CloudGraphsDFG)::Nothing
+function clearUser!!(dfg::Neo4jDFG)
     # Perform detach+deletion
     _queryNeo4j(dfg.neo4jInstance, "match (node:$(dfg.userId)) detach delete node ")
 
@@ -268,8 +315,11 @@ end
     $(SIGNATURES)
 DANGER: Copies and overwrites the destination session.
 If no destination specified then it creates a unique one.
+
+Notes
+- Returns `::Neo4jDFG `
 """
-function copySession!(sourceDFG::CloudGraphsDFG, destDFG::Union{Nothing, CloudGraphsDFG})::CloudGraphsDFG
+function copySession!(sourceDFG::Neo4jDFG, destDFG::Union{Nothing, <:Neo4jDFG})
     if destDFG == nothing
         destDFG = _getDuplicatedEmptyDFG(sourceDFG)
     end
@@ -279,5 +329,8 @@ end
 """
     $(SIGNATURES)
 DANGER: Copies the source to a new unique destination.
+
+Notes
+- Returns `::Neo4jDFG`
 """
-copySession!(sourceDFG::CloudGraphsDFG)::CloudGraphsDFG = copySession!(sourceDFG, nothing)
+copySession!(sourceDFG::Neo4jDFG) = copySession!(sourceDFG, nothing)